Split progress source tests
